Split drop-delegate types out of SidebarWorkspaceGroupHeaderView
The added menu items plus the upstream hover-tracker rework pushed
SidebarWorkspaceGroupHeaderView.swift to 560 lines, over the 500-line
file-length budget (workflow-guard-tests). Move the cohesive, separable
DropZone/DropAction/DropPolicy/DropDelegate types into a new
SidebarWorkspaceGroupHeaderDropDelegate.swift (wired into the pbxproj and
normalized), leaving the view file at 372 lines. No behavior change.
